requirements_frontend.txt
--extra-index-url https://download.pytorch.org/whl/cu121
python-dotenv==1.0.1
fastapi==0.112.2
fonttools==4.53.1
gradio==4.38.1
gradio_client== 1.1.0
httpx==0.27.2
huggingface-hub==0.23.4
lancedb==0.10.0
langchain==0.2.8
langchain-community==0.2.1
langchain-core==0.2.19
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
matplotlib==3.9.1
onnx==1.17.0
openai==1.35.14
pillow==10.4.0
pillow_heif==0.17.0
pymupdf==1.24.7
pypdf==4.3.0
sentence-transformers==3.0.1
torch==2.3.0
transformers==4.42.4
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.14.10
unstructured-client==0.24.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12